+++ Source/Krypton Toolkit/Krypton.Toolkit.Suite.Extended.Core/Classes/UAC/UACUtilityMethods.cs
@@ -1,21 +1,13 @@
-﻿#region BSD License
-/*
- * Use of this source code is governed by a BSD-style
- * license or other governing licenses that can be found in the LICENSE.md file or at
- * https://raw.githubusercontent.com/Krypton-Suite/Extended-Toolkit/master/LICENSE
- */
-#endregion
-
-namespace Krypton.Toolkit.Suite.Extended.Buttons
+﻿namespace Krypton.Toolkit.Suite.Extended.Core
 {
-    /// <summary>Provides useful methods for specified tasks.</summary>
-    public class UtilityMethods
+    /// <summary>Utility methods for the user account control.</summary>
+    public class UACUtilityMethods
     {
         #region Constructor
         /// <summary>
-        /// Initialises a new instance of <see cref="UtilityMethods"/>.
+        /// Initialises a new instance of <see cref="UACUtilityMethods"/>.
         /// </summary>
-        public UtilityMethods()
+        public UACUtilityMethods()
         {
 
         }
@@ -23,7 +15,7 @@ namespace Krypton.Toolkit.Suite.Extended.Buttons
 
         #region Methods
         /// <summary>
-        /// Elevates the application to use administrative privileges. To be used with <see cref="Krypton.Toolkit.Suite.Extended.Buttons.KryptonUACElevatedButton"/> or <see cref="Krypton.Toolkit.Suite.Extended.Tool.Strip.Items.ToolStripMenuItemUACSheld"/> button click.
+        /// Elevates the application to use administrative privileges. To be used with <see cref="KryptonUACButtonVersion1"/> or <see cref="ToolStripMenuItemUACSheld"/> button click.
         /// </summary>
         /// <param name="processName">The process name that you wish to elevate.</param>
         public static void ElevateProcessWithAdministrativeRights(string processName)
@@ -47,7 +39,7 @@ namespace Krypton.Toolkit.Suite.Extended.Buttons
                 }
                 catch (Win32Exception wexc)
                 {
-                    KryptonMessageBox.Show("Error: " + wexc.Message, "An Error has Occurred", MessageBoxButtons.OK, MessageBoxIcon.Error);
+                    ExceptionCapture.CaptureException(wexc);
                 }
 
                 return;