+++ Source/Krypton Toolkit/Krypton.Toolkit.Suite.Extended.Data.Visualisation/Globals/GlobalDeclarations.cs
@@ -8,13 +8,11 @@
 
 global using System;
 global using System.ComponentModel;
-global using System.ComponentModel.Design;
 global using System.Collections.Generic;
 global using System.Collections.ObjectModel;
 global using System.Collections.Specialized;
 global using System.Diagnostics;
 global using System.Drawing;
-global using System.Drawing.Design;
 global using System.Drawing.Drawing2D;
 global using System.Drawing.Text;
 global using System.Drawing.Imaging;
@@ -28,11 +26,3 @@ global using System.Text;
 global using System.Threading.Tasks;
 global using System.Linq;
 global using System.Windows.Forms;
-global using System.Windows.Forms.Design;
-global using System.Windows.Forms.VisualStyles;
-
-
-global using Microsoft.Win32;
-
-global using Krypton.Toolkit.Suite.Extended.Data.Visualisation;
-global using Krypton.Toolkit.Suite.Extended.Data.Visualisation.ScottPlot.Drawing;