+++ Source/Krypton Toolkit/Krypton.Toolkit.Suite.Extended.Dialogs/Globals/GlobalDeclarations.cs
@@ -18,9 +18,7 @@ global using System.Text;
 global using System.Windows.Forms;
 
 global using Krypton.Toolkit.Suite.Extended.Common;
-global using Krypton.Toolkit.Suite.Extended.Controls;
 global using Krypton.Toolkit.Suite.Extended.Developer.Utilities;
-global using Krypton.Toolkit.Suite.Extended.Dialogs.Properties;
 global using Krypton.Toolkit.Suite.Extended.Global.Utilities;
 global using Krypton.Toolkit.Suite.Extended.Settings;
 global using Krypton.Toolkit.Suite.Extended.Utilities.System.Synthesis;