+++ Source/Krypton Toolkit/Krypton.Toolkit.Suite.Extended.Drawing.Utilities/Globals/GlobalDeclarations.cs
@@ -30,7 +30,5 @@ global using Cyotek.Windows.Forms;
 global using Krypton.Toolkit.Suite.Extended.Buttons;
 global using Krypton.Toolkit.Suite.Extended.Core;
 global using Krypton.Toolkit.Suite.Extended.Common;
-global using Krypton.Toolkit.Suite.Extended.Controls;
-global using Krypton.Toolkit.Suite.Extended.Drawing.Utilities.Properties;
 global using Krypton.Toolkit.Suite.Extended.Developer.Utilities;
 global using Krypton.Toolkit.Suite.Extended.Messagebox;