+++ Source/Krypton Toolkit/Krypton.Toolkit.Suite.Extended.Memory.Box/Globals/GlobalDeclarations.cs
@@ -11,4 +11,3 @@ global using System.Drawing;
 global using System.Windows.Forms;
 
 global using Krypton.Toolkit.Suite.Extended.Common;
-global using Krypton.Toolkit.Suite.Extended.Memory.Box.Properties;