+++ Source/Krypton Toolkit/Krypton.Toolkit.Suite.Extended.Messagebox/Globals/GlobalDeclarations.cs
@@ -14,5 +14,4 @@ global using System.Media;
 global using System.Runtime.InteropServices;
 global using System.Reflection;
 global using System.Text;
-global using System.Threading;
 global using System.Windows.Forms;